security: remove the security_netlink_recv hook as it is equivalent to capable()
Once upon a time netlink was not sync and we had to get the effective
capabilities from the skb that was being received.  Today we instead get
the capabilities from the current task.  This has rendered the entire
purpose of the hook moot as it is now functionally equivalent to the
capable() call.

Change-Id: Iaeca0a350334256119d0359921810c26923649ff
Signed-off-by: Eric Paris <eparis@redhat.com>