proc: uid: Adds accounting for the cputimes per uid.
Adds proc files /proc/uid_cputime/show_uid_stat and
/proc/uid_cputime/remove_uid_range.

show_uid_stat lists the total utime and stime for the active as well as
terminated processes for each of the uids.

Writing a range of uids to remove_uid_range will delete the accounting
for all the uids within that range.

Conflicts:
	drivers/misc/Kconfig
	drivers/misc/Makefile

Change-Id: Ifbdd2bcd218bd9e97a570eeaf7f428ebf79e369b